% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_validation.R
\name{import_validation}
\alias{import_validation}
\title{Check column headers for compatibility}
\usage{
import_validation(dat1, dat2)
}
\arguments{
\item{dat1}{A data frame of adult attendance at each nest of interest. The data from should have the followng headers (case senstive):SPLIT_YEAR,    ROOKERY,COLONY,    CAMERA,    SPP, NEST,    DATE (m/d/y), and    MAXN}

\item{dat2}{A data frame of nest content observations. The data frame shoulhd have the following headers (case senstive):SPLIT_YEAR,  ROOKERY,  SPP,  COLONY,  CAMERA,  NEST,}
}
\value{
A list of errors in data headings
}
\description{
Check column headers for compatibility
}
