% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology_app.R
\name{phenology_app}
\alias{phenology_app}
\title{phenology_app}
\usage{
phenology_app(dat, rdat, tabletype = "Raw", degf = 10)
}
\arguments{
\item{dat}{A data frame of adult attendance at each nest of interest. The data should have the following headers (case sensitive):SPLIT_YEAR, ROOKERY, COLONY, CAMERA,    SPP, NEST,    DATE, and    MAXN. Note that DATE can be specified in any way (default is for m/d/y), or as separate columns named DAY, MON, YR.}

\item{rdat}{A data frame of nest content observations. The data frame should have the following headers (case sensitive):SPLIT_YEAR,  ROOKERY,  SPP,  COLONY,  CAMERA,  NEST, DATE,  LAY,  MAXE,  HATCH,  MAXC, and  CRECHE. Note that DATE can be specified in anyway (default is for m/d/y), or as separate columns named DAY, MON, YR.}

\item{tabletype}{A character vector indicating output of interest, either "Raw", "CEMP Protocol A6b", or "CEMP Protocol A9"}

\item{degf}{numeric value for degrees of freedom used in the smooth.spline function}
}
\value{
A data frame with results
}
\description{
Function to process adult attendance of nest data and nest content data to estimate phenology and reproductive success. The app requests you input the data as a .csv file. To run this file independently, you will need to import the .csv and format the date field to POSIXct appropriately.
}
\examples{
phenology_app(test_att, test_repro, tabletype="Raw", degf=10)
}
