% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_att}
\alias{test_att}
\title{test_att}
\format{
A dataframe with of adult attendance at each nest of interest. The data should have the following headers (case sensitive):SPLIT_YEAR, ROOKERY, COLONY, CAMERA,	SPP, NEST,	DATE, and	MAXN. Note that DATE can be specified in any way (default is for m/d/y), or as separate columns named DAY, MON, YR.
}
\usage{
test_att
}
\description{
An example data set for running auxilliary functions of the Shiny app. The data in test_att are nest attendance data that reocrd teh presense of adult birds at their nest on a daily basis.
}
\keyword{datasets}
