% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_repro}
\alias{test_repro}
\title{test_repro}
\format{
A data frame of nest content observations. The data frame should have the following headers (case sensitive):SPLIT_YEAR,  ROOKERY,  SPP,  COLONY,  CAMERA,  NEST, DATE,  LAY,  MAXE,  HATCH,  MAXC, and  CRECHE. Note that DATE can be specified in anyway (default is for m/d/y), or as separate columns named DAY, MON, YR.
}
\usage{
test_repro
}
\description{
An example data set for running auxilliary functions of the Shiny app. The data in test_repro are nest content observations on a daily basis.
}
\keyword{datasets}
