% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_checking.R
\name{error_checking}
\alias{error_checking}
\title{eerror_checking}
\usage{
error_checking(dat1 = dat, dat2 = rdat)
}
\arguments{
\item{dat1}{A data frame of adult attendance at each nest of interest. The data from should have the followng headers (case senstive):SPLIT_YEAR,    ROOKERY,COLONY,    CAMERA,    SPP, NEST,    DATE (m/d/y), and    MAXN}

\item{dat2}{A data frame of nest content observations. The data frame should have the following headers (case senstive):SPLIT_YEAR,  ROOKERY,  SPP,  COLONY,  CAMERA,  NEST,}
}
\value{
A list with information about the nature and location (row) of suspected errors in each data file
}
\description{
eerror_checking
}
