% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_checking.R
\name{error_checking}
\alias{error_checking}
\title{error_checking}
\usage{
error_checking(dat1, dat2)
}
\arguments{
\item{dat1}{A data frame of adult attendance at each nest of interest. The data from should have the following headers (case sensitive):SPLIT_YEAR, ROOKERY,COLONY, CAMERA,    SPP, NEST,    DATE, and    MAXN. Note that DATE can be specified in anyway (default is for m/d/y), or as separate columns named DAY, MON, YR.}

\item{dat2}{A data frame of nest content observations. The data frame should have the following headers (case sensitive):SPLIT_YEAR,  ROOKERY,  SPP,  COLONY,  CAMERA,  NEST, DATE,  LAY,  MAXE,  HATCH,  MAXC, and  CRECHE. Note that DATE can be specified in anyway (default is for m/d/y), or as separate columns named DAY, MON, YR.}
}
\value{
A list with information about the nature and location (row) of suspected errors in each data file. If no errors are encounterd, list has length 0.
}
\description{
This function checks the input data frames for common entry errors and returns information to support the analyst in cleaning the data files if errors are encountered.This runs in the background and you shouldn't need to access this function. To run this file independently, you will need to import the .csv and format the date field to POSIXct appropriately.
}
\examples{
error_checking(test_att, test_repro)
}
