% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_validation.R
\name{import_validation}
\alias{import_validation}
\title{import_validation}
\usage{
import_validation(dat, rdat)
}
\arguments{
\item{dat}{A data frame of adult attendance at each nest of interest. The data from should have the following headers (case sensitive):SPLIT_YEAR, ROOKERY,COLONY, CAMERA,    SPP, NEST,    DATE, and    MAXN. Note that DATE can be specified in anyway (default is for m/d/y), or as separate columns named DAY, MON, YR.}

\item{rdat}{A data frame of nest content observations. The data frame should have the following headers (case sensitive):SPLIT_YEAR,  ROOKERY,  SPP,  COLONY,  CAMERA,  NEST, DATE,  LAY,  MAXE,  HATCH,  MAXC, and  CRECHE. Note that DATE can be specified in anyway (default is for m/d/y), or as separate columns named DAY, MON, YR.}
}
\value{
A list of errors in data headings. If no errors, list has length 0.
}
\description{
Function to check input data headers for compatibility with the app. Returns information if required headers are missing. This runs in the background and the user should not need to access this function. To run this file independently, you will need to import your .csv files and ensure the 'DATE" field is formatted as POSIXct.
}
\examples{
import_validation(test_att, test_repro)
}
